+++ examples/ff.rs
@@ -8,7 +8,7 @@ use std::time::Duration;
 
 fn main() {
     env_logger::init().unwrap();
-    let mut gilrs = Gilrs::new();
+    let mut gilrs = Gilrs::new().unwrap();
     let support_ff = gilrs
         .gamepads()
         .filter_map(|(id, gp)| if gp.is_ff_supported() { Some(id) } else { None })